requirements.txt
tensorflow>=2.10
opencv-python
pillow
pandas
scikit-learn
streamlit
tqdm
tensorflow>=2.12
numpy

